main activity moved to fragment

+++ app/src/main/java/com/example/android/bakingapp/model/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient implements Parcelable
 
     @SerializedName("quantity")
     @Expose
-    private int quantity;
+    private double quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -36,7 +36,7 @@ public class Ingredient implements Parcelable
             ;
 
     protected Ingredient(Parcel in) {
-        this.quantity = ((int) in.readValue((int.class.getClassLoader())));
+        this.quantity = ((double) in.readValue((double.class.getClassLoader())));
         this.measure = ((String) in.readValue((String.class.getClassLoader())));
         this.ingredient = ((String) in.readValue((String.class.getClassLoader())));
     }
@@ -54,18 +54,18 @@ public class Ingredient implements Parcelable
      * @param ingredient
      * @param quantity
      */
-    public Ingredient(int quantity, String measure, String ingredient) {
+    public Ingredient(double quantity, String measure, String ingredient) {
         super();
         this.quantity = quantity;
         this.measure = measure;
         this.ingredient = ingredient;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
